refractor: move controls folder as a subfolder of visualizer
Rename volume-renderer folder as visualizer and add Visualizer.js
