+++ src/components/React/visualizer/Visualizer.js
@@ -1,8 +1,8 @@
 import { Container } from 'react-bootstrap';
 
-import VolumeRenderer from './volume-renderer/VolumeRenderer';
+import VolumeRenderer from './VolumeRenderer';
 import ControlsPanel from './control-panel/ControlsPanel';
-import { VolumeProvider } from '../../context/volume-context';
+import { VolumeProvider } from '../../../context/volume-context';
 
 export default function Visualizer(props) {
   return (
