+++ src/components/React/visualizer/control-panel/OpacityControl.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react'
 import { connect } from "react-redux";
-import {mySendAlphaPoints} from '../../../redux/AppActions'
+import {mySendAlphaPoints} from '../../../../redux/AppActions'
 
 const mapStateToProps = state => {
   return {volumeData: state.volumeData};
